+++ routes/ui.php
@@ -2,6 +2,7 @@
 
 use Illuminate\Support\Facades\Route;
 use Enadstack\LaravelRoles\Http\Controllers\UI\RoleUIController;
+use Enadstack\LaravelRoles\Http\Controllers\UI\PermissionUIController;
 use Enadstack\LaravelRoles\Http\Controllers\UI\MatrixUIController;
 
 /*
@@ -14,6 +15,7 @@ use Enadstack\LaravelRoles\Http\Controllers\UI\MatrixUIController;
 |   AND config('roles.ui.driver') === 'vue'
 |
 | These are Inertia routes that render Vue pages.
+| All routes use config-based prefixes - no hardcoded URLs.
 |
 */
 
@@ -33,6 +35,16 @@ Route::middleware($middleware)
         Route::get('/roles/create', [RoleUIController::class, 'create'])->name('roles.create');
         Route::get('/roles/{id}', [RoleUIController::class, 'show'])->name('roles.show');
         Route::get('/roles/{id}/edit', [RoleUIController::class, 'edit'])->name('roles.edit');
+
+        /*
+        |--------------------------------------------------------------------------
+        | Permissions UI Pages
+        |--------------------------------------------------------------------------
+        */
+        Route::get('/permissions', [PermissionUIController::class, 'index'])->name('permissions.index');
+        Route::get('/permissions/create', [PermissionUIController::class, 'create'])->name('permissions.create');
+        Route::get('/permissions/{id}', [PermissionUIController::class, 'show'])->name('permissions.show');
+        Route::get('/permissions/{id}/edit', [PermissionUIController::class, 'edit'])->name('permissions.edit');
 
         /*
         |--------------------------------------------------------------------------